+++ TransactionalBox.sln
@@ -23,10 +23,6 @@ Project("{E53339B2-1760-4266-BCC7-CA923CBCF16C}") = "TransactionalBoxWithWorkers
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TransactionalBox.Sample.WebApi.InMemory", "samples\TransactionalBox.Sample.WebApi.InMemory\TransactionalBox.Sample.WebApi.InMemory.csproj", "{5C2EB157-F976-42F9-8B26-42A50898714C}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Internals", "Internals", "{9C70A8B9-322F-49D5-AF08-889B6BBB161C}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TransactionalBox.Internals.EventHooks", "source\Internals\TransactionalBox.Internals.EventHooks\TransactionalBox.Internals.EventHooks.csproj", "{891F6713-08CE-4FCF-881C-BBA258C18C72}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TransactionalBox.End2EndTests", "tests\TransactionalBox.End2EndTests\TransactionalBox.End2EndTests.csproj", "{06A55E8C-9D5B-4E12-B64B-EC59690818C9}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Bank", "Bank", "{6BB61EDF-B793-4CD2-94F5-9B11E131845E}"
@@ -83,10 +79,6 @@ Global
 		{5C2EB157-F976-42F9-8B26-42A50898714C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5C2EB157-F976-42F9-8B26-42A50898714C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5C2EB157-F976-42F9-8B26-42A50898714C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{891F6713-08CE-4FCF-881C-BBA258C18C72}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{891F6713-08CE-4FCF-881C-BBA258C18C72}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{891F6713-08CE-4FCF-881C-BBA258C18C72}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{891F6713-08CE-4FCF-881C-BBA258C18C72}.Release|Any CPU.Build.0 = Release|Any CPU
 		{06A55E8C-9D5B-4E12-B64B-EC59690818C9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{06A55E8C-9D5B-4E12-B64B-EC59690818C9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{06A55E8C-9D5B-4E12-B64B-EC59690818C9}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -135,8 +127,6 @@ Global
 		{26ECF009-798E-4AEA-B8B1-E11FEED33FB0} = {D018E603-5AE9-4B1B-BF33-04EB0B5D6913}
 		{D89EDC24-3316-4A91-8442-6D1ABAFEDEBF} = {D018E603-5AE9-4B1B-BF33-04EB0B5D6913}
 		{5C2EB157-F976-42F9-8B26-42A50898714C} = {D018E603-5AE9-4B1B-BF33-04EB0B5D6913}
-		{9C70A8B9-322F-49D5-AF08-889B6BBB161C} = {8FC46C4A-3C0C-458F-ACF5-F35F728A2988}
-		{891F6713-08CE-4FCF-881C-BBA258C18C72} = {9C70A8B9-322F-49D5-AF08-889B6BBB161C}
 		{06A55E8C-9D5B-4E12-B64B-EC59690818C9} = {6DEFBC37-33F1-4E42-9A7E-2752E9AD3187}
 		{6BB61EDF-B793-4CD2-94F5-9B11E131845E} = {D018E603-5AE9-4B1B-BF33-04EB0B5D6913}
 		{580972A5-B9DC-49B0-8CFB-28412E3FC5C5} = {6BB61EDF-B793-4CD2-94F5-9B11E131845E}
