--- source/Internals/TransactionalBox.Internals.EventHooks/EventHook.cs
@@ -1,4 +1,0 @@
-﻿namespace TransactionalBox.Internals.EventHooks
-{
-    internal abstract class EventHook;
-}
