--- source/Internals/TransactionalBox.Internals.EventHooks/TransactionalBox.Internals.EventHooks.csproj
@@ -1,11 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-	<ItemGroup>
-		<InternalsVisibleTo Include="TransactionalBox" />
-		<InternalsVisibleTo Include="TransactionalBox.Inbox" />
-	</ItemGroup>
-
-	<ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" />
-    <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" />
-  </ItemGroup>
-</Project>
