+++ source/TransactionalBox/Internals/Inbox/Hooks/Handlers/CleanUpInbox/CleanUpInbox.cs
@@ -1,8 +1,7 @@
-﻿using TransactionalBox.Internals.EventHooks;
-using TransactionalBox.Internals.EventHooks.Contexts;
-using TransactionalBox.Internals.Inbox.Hooks.Handlers.CleanUpInbox.Logger;
+﻿using TransactionalBox.Internals.Inbox.Hooks.Handlers.CleanUpInbox.Logger;
 using TransactionalBox.Internals.Inbox.Storage.ContractsToImplement;
 using TransactionalBox.Internals.Inbox.Hooks.Events;
+using TransactionalBox.Internals.InternalPackages.EventHooks;
 
 namespace TransactionalBox.Internals.Inbox.Hooks.Handlers.CleanUpInbox
 {
