+++ source/TransactionalBox/Internals/InternalPackages/EventHooks/EventHook.cs
@@ -1,0 +1,4 @@
+﻿namespace TransactionalBox.Internals.InternalPackages.EventHooks
+{
+    internal abstract class EventHook;
+}
