+++ source/TransactionalBox/Internals/InternalPackages/EventHooks/HookExecutionContext.cs
@@ -1,6 +1,4 @@
-﻿using TransactionalBox.Internals.EventHooks.Contexts;
-
-namespace TransactionalBox.Internals.EventHooks.Internals.Contexts
+﻿namespace TransactionalBox.Internals.InternalPackages.EventHooks
 {
     internal sealed record HookExecutionContext(Guid Id, string Name, DateTime LastOccurredUtc, bool IsError, long Attempt) : IHookExecutionContext;
 }
